fix(dashboard): move the platform layer helper out of the page module
