+++ client/dashboard/src/pages/device-agent/device-agent-configuration.tsx
@@ -28,18 +28,14 @@ import {
 import { useUpdateDeviceAgentConfigurationMutation } from "@gram/client/react-query/updateDeviceAgentConfiguration.js";
 import { Button } from "@/components/ui/Button";
 import { Stack } from "@/components/ui/Stack";
+import {
+  enforcementLayer,
+  recordValue,
+  type EnforcementLayer,
+} from "./platform-layers";
 import { useQueryClient } from "@tanstack/react-query";
 import { type ReactNode, useState } from "react";
 import { toast } from "sonner";
-
-type EnforcementLayer = "off" | "user" | "managed";
-
-// The `platforms` map is opt-out on the device agent: a tool absent from it is
-// managed at the user layer (config.PlatformMode.Layer in the device-agent
-// repo). Only an explicit `false` disables one. This UI must resolve an absent
-// key the same way, or it reports a tool as Off while the fleet is enforcing
-// it — and then writes that false reading back on the next save.
-const DEFAULT_LAYER: EnforcementLayer = "user";
 
 const PLATFORMS = [
   {
@@ -94,23 +90,6 @@ const AUTO_UPDATE_POLICIES = [
     description: "Install updates as they release",
   },
 ] as const;
-
-function recordValue(value: unknown): Record<string, unknown> {
-  if (typeof value === "object" && value !== null && !Array.isArray(value)) {
-    return value as Record<string, unknown>;
-  }
-  return {};
-}
-
-export function enforcementLayer(
-  platforms: unknown,
-  platformKey: string,
-): EnforcementLayer {
-  const value = recordValue(platforms)[platformKey];
-  if (value === "user" || value === "managed") return value;
-  if (value === false) return "off";
-  return DEFAULT_LAYER;
-}
 
 function stringSetting(
   config: DeviceAgentConfiguration,
